requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
djangorestframework==3.16.0
requests==2.32.3
rocrate==0.13.0
